requirements.txt
click==8.1.7
colorama==0.4.6
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==1.26.4
pandas==2.2.1
Pygments==2.17.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
rich==13.7.1
six==1.16.0
typer==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
